+++ src/js/core/accordion.js
@@ -46,12 +46,10 @@ export default {
     computed: {
         items: {
             get({ targets }, $el) {
-                return $$(targets, $el).filter((el) => $(this.content, el));
+                return $$(targets, $el);
             },
 
             watch(items, prev) {
-                items.forEach((el) => hide($(this.content, el), !hasClass(el, this.clsOpen)));
-
                 if (prev || hasClass(items, this.clsOpen)) {
                     return;
                 }
@@ -70,6 +68,26 @@ export default {
 
         toggles({ toggle }) {
             return this.items.map((item) => $(toggle, item));
+        },
+
+        contents: {
+            get({ content }) {
+                return this.items.map((item) => $(content, item));
+            },
+
+            watch(items) {
+                for (const el of items) {
+                    hide(
+                        el,
+                        !hasClass(
+                            this.items.find((item) => item.contains(el)),
+                            this.clsOpen
+                        )
+                    );
+                }
+            },
+
+            immediate: true,
         },
     },