+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fix toggling accordion item while transition is in progress
 - Fix Drop positioning with `boundaryAlign: true`
 - Fix Scrollspy component not triggering in some cases
+- Fix Slider component showing wrong initial index
 
 ## 3.14.0 (April 27, 2022)